Move command line parser to UI #33
- changed resources path name following global namespace change
- tested using BuildStandardTypes.bat
