+++ ModelCompilerUI/ModelCompilerAPIInternal.cs
@@ -5,10 +5,11 @@
 //  To be in touch join the community at GitHub: https://github.com/mpostol/OPC-UA-OOI/discussions
 //__________________________________________________________________________________________________
 
+using OOI.ModelCompiler;
 using System;
 using System.Collections.Generic;
 
-namespace OOI.ModelCompiler
+namespace OOI.ModelCompilerUI
 {
   public class ModelCompilerAPIInternal : ModelCompilerAPI
   {
@@ -19,9 +20,7 @@ namespace OOI.ModelCompiler
         if (tokens[ii] == "-input")
         {
           if (ii >= tokens.Count - 1)
-          {
             throw new ArgumentException("Incorrect number of parameters specified with the -input option.");
-          }
           inputDirectory = tokens[++ii];
           continue;
         }
@@ -42,10 +41,7 @@ namespace OOI.ModelCompiler
         if (tokens[ii] == "-license")
         {
           if (ii >= tokens.Count - 1)
-          {
             throw new ArgumentException("Incorrect number of parameters specified with the -license option.");
-          }
-
           updateHeaders = true;
           licenseType = (LicenseType)Enum.Parse(typeof(LicenseType), tokens[++ii]);
           continue;
@@ -53,30 +49,21 @@ namespace OOI.ModelCompiler
         if (tokens[ii] == "-d2")
         {
           if (ii >= tokens.Count - 1)
-          {
             throw new ArgumentException("Incorrect number of parameters specified with the -d2 option.");
-          }
-
           designFiles.Add(tokens[++ii]);
           continue;
         }
         if (tokens[ii] == "-d2")
         {
           if (ii >= tokens.Count - 1)
-          {
             throw new ArgumentException("Incorrect number of parameters specified with the -d2 option.");
-          }
-
           designFiles.Add(tokens[++ii]);
           continue;
         }
         if (tokens[ii] == "-d2")
         {
           if (ii >= tokens.Count - 1)
-          {
             throw new ArgumentException("Incorrect number of parameters specified with the -d2 option.");
-          }
-
           designFiles.Add(tokens[++ii]);
           continue;
         }
@@ -86,7 +73,6 @@ namespace OOI.ModelCompiler
           {
             throw new ArgumentException("Incorrect number of parameters specified with the -c or -cg option.");
           }
-
           generateIds = tokens[ii] == "-cg";
           identifierFile = tokens[++ii];
           continue;
@@ -94,20 +80,14 @@ namespace OOI.ModelCompiler
         if (tokens[ii] == "-o")
         {
           if (ii >= tokens.Count - 1)
-          {
             throw new ArgumentException("Incorrect number of parameters specified with the -o option.");
-          }
-
           outputDir = tokens[++ii];
           continue;
         }
         if (tokens[ii] == "-o2")
         {
           if (ii >= tokens.Count - 1)
-          {
             throw new ArgumentException("Incorrect number of parameters specified with the -o option.");
-          }
-
           outputDir = tokens[++ii];
           generateMultiFile = true;
           continue;
@@ -120,20 +100,14 @@ namespace OOI.ModelCompiler
         if (tokens[ii] == "-version")
         {
           if (ii >= tokens.Count - 1)
-          {
             throw new ArgumentException("Incorrect number of parameters specified with the -version option.");
-          }
-
           specificationVersion = tokens[++ii];
           continue;
         }
         if (tokens[ii] == "-ansic")
         {
           if (ii >= tokens.Count - 1)
-          {
             throw new ArgumentException("Incorrect number of parameters specified with the -ansic option.");
-          }
-
           ansicRootDir = tokens[++ii];
           continue;
         }
@@ -159,10 +133,7 @@ namespace OOI.ModelCompiler
         if (tokens[ii] == "-exclude")
         {
           if (ii >= tokens.Count - 1)
-          {
             throw new ArgumentException("Incorrect number of parameters specified with the -exclude option.");
-          }
-
           excludeCategories = tokens[++ii].Split(',', '+');
           continue;
         }
