+++ ModelCompilerUI/ModelCompilerUI.csproj
@@ -108,6 +108,7 @@
       <DependentUpon>ExceptionDlg.cs</DependentUpon>
     </Compile>
     <Compile Include="MeasurementUnits.cs" />
+    <Compile Include="ModelCompilerAPIInternal.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ToForms\GUIHandling.cs" />
     <Compile Include="ToForms\IGUIHandling.cs" />
